Add files via upload
Created a web page using the following: XML, XSLT, and CSS
The XML document will contain the following information for the page:
 Title (for the web page)
 Menu (These should link to some news sites, not the actual articles.)
o Link text (This will be surrounded by an <a> tag with the corresponding href)
o href (the actual link displayed on the page; e.g., http://www.mynews.net/news), this link should NOT be clickable.
 Content
o paragraphs of content
 News
o Title (This should also be a link that goes to the full news article.)
o Intro (short snippet of the news article)
o href (the displayed actual link to the article; e.g., http://www.mynews.net/news), again, this should not be clickable.
o Date of article
For the news articles, have them sorted by date. How you layout the page is up to you but make sure you create your XML so that your .xsl stylesheet can properly render
the HTML output. Use CSS to specify fonts, text formatting, and colours in order to make it look presentable; you will be marked on the appearance (subjective) of your page.
(Note: No namespaces are required in your source XML file for this project.)